Innovorex LLM/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Vector Storage
faiss-cpu==1.7.4
numpy==1.24.3

# ML and NLP
torch==2.1.0+cpu
transformers==4.36.0
sentence-transformers==2.2.2
datasets==2.14.6
peft==0.7.1
accelerate==0.24.1

# External APIs
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.2
httpx==0.25.2

# Security and Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
aiofiles==23.2.1
jinja2==3.1.2

# System monitoring
psutil==5.9.6

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0